+++ subsys/logging/backends/log_backend_net.c
@@ -367,7 +367,7 @@ const struct log_backend *log_backend_net_get(void)
 }
 
 #if defined(CONFIG_LOG_BACKEND_NET_USE_CONNECTION_MANAGER)
-static void l4_event_handler(uint32_t mgmt_event, struct net_if *iface, void *info,
+static void l4_event_handler(uint64_t mgmt_event, struct net_if *iface, void *info,
 			     size_t info_length, void *user_data)
 {
 	ARG_UNUSED(iface);
